requirements.txt
langchain
langchain-community
langchain-openai
openai
python-dotenv
supabase
streamlit
pip install langchain langchain-community faiss-cpu
faiss-cpu
